Struct tor_rtmock::net::MockNetProvider

pub struct MockNetProvider { /* private fields */ }

A view of a single host’s access to a MockNetwork.

Each simulated host has its own addresses that it’s allowed to listen on, and a reference to the network.

This type implements TcpProvider so that it can be used as a drop-in replacement for testing code that uses the network.

Limitations

There’s no randomness here, so we can’t simulate the weirdness of real networks.

So far, there’s no support for DNS or UDP.

We don’t handle localhost specially, and we don’t simulate providers that can connect to some addresses but not all.

We don’t do the right thing (block) if there is a listener that never calls accept.

UDP is completely broken: datagrams appear to be transmitted, but will never be received. And local address assignment is not implemented so .local_addr() can return NONE

We use a simple u16 counter to decide what arbitrary port numbers to use: Once that counter is exhausted, we will fail with an assertion. We don’t do anything to prevent those arbitrary ports from colliding with specified ports, other than declare that you can’t have two listeners on the same addr:port at the same time.

We pretend to provide TLS, but there’s no actual encryption or authentication.

Implementations

impl MockNetProvider

pub fn listen_tls( &self, addr: &SocketAddr, tls_cert: Vec<u8> ) -> IoResult<MockNetListener>

Create a mock TLS listener with provided certificate.

Note that no encryption or authentication is actually performed! Other parties are simply told that their connections succeeded and were authenticated against the given certificate.