Struct tor_rtmock::MockRuntime

pub struct MockRuntime { /* private fields */ }

Completely mock runtime

Suitable for test cases that wish to completely control the environment experienced by the code under test.

Restrictions

The test case must advance the mock time explicitly as desired. There is not currently any facility for automatically making progress by advancing the mock time by the right amounts for the timeouts set by the futures under test.

Tests that use this runtime must not interact with the outside world; everything must go through this runtime (and its pieces).

Allowed

Inter-future communication facilities from futures or other runtime-agnostic crates.

Fast synchronous operations that will complete “immediately” or “quickly”. E.g.: filesystem calls.

std::sync::Mutex (assuming the use is deadlock-free in a single-threaded executor, as it should be in all of Arti).

Slower operations that are run synchronously (without futures await) provided their completion doesn’t depend on any of the futures we’re running. (These kind of operations are often discouraged in async contexts, because they block the async runtime or its worker threads. But they are often OK in tests.)

All facilities provided by this MockExecutor and its trait impls.

Not allowed

Direct access to the real-world clock (SystemTime::now, Instant::now). Including coarsetime, which is not mocked. Exception: CPU use measurements.

Anything that spawns threads and then communicates with those threads using async Rust facilities (futures).

Async sockets, or async use of other kernel-based IPC or network mechanisms.

Anything provided by a Rust runtime/executor project (eg anything from Tokio), unless it is definitively established that it’s runtime-agnostic.

Implementations

impl MockRuntime

pub fn mock_task(&self) -> &MockExecutor

Method borrowing MockRuntime::task field. Tasks

pub fn mock_sleep(&self) -> &MockSleepProvider

Method borrowing MockRuntime::sleep field. Time provider

pub fn mock_net(&self) -> &MockNetProvider

Method borrowing MockRuntime::net field. Net provider

impl MockRuntime

pub fn new() -> Self

Create a new MockRuntime with default parameters

pub fn builder() -> MockRuntimeBuilder

Return a builder, for creating a MockRuntime with some parameters manually configured

pub fn test_with_various<TC, FUT>(test_case: TC)where TC: FnMut(MockRuntime) -> FUT, FUT: Future<Output = ()>,

Run a test case with a variety of runtime parameters, to try to find bugs

test_case is an async closure which receives a MockRuntime. It will be run with a number of differently configured executors.

Variations

The only variation currently implemented is this:

Both FIFO and LIFO scheduling policies are tested, in the hope that this will help discover ordering-dependent bugs.

pub fn try_test_with_various<TC, FUT, E>(test_case: TC) -> Result<(), E>where TC: FnMut(MockRuntime) -> FUT, FUT: Future<Output = Result<(), E>>,

Run a faillible test case with a variety of runtime parameters, to try to find bugs

test_case is an async closure which receives a MockRuntime. It will be run with a number of differently configured executors.

This function accepts a fallible closure, and returns the first Err to the caller.

See test_with_various() for more details.

pub async fn progress_until_stalled(&self)

Run tasks in the current executor until every task except this one is waiting

Calls MockExecutor::progress_until_stalled().

Restriction - no automatic time advance

The mocked time will not be automatically advanced.

Usually (and especially if the tasks under test are waiting for timeouts or periodic events) you must use advance() or jump_to() to ensure the simulated time progresses as required.

Panics

Might malfunction or panic if more than one such call is running at once.

(Ie, you must .await or drop the returned Future before calling this method again.)

Must be called and awaited within a future being run by self.

pub async fn advance(&self, dur: Duration)

See MockSleepProvider::advance()

pub fn jump_to(&self, new_wallclock: SystemTime)

See MockSleepProvider::jump_to()