Struct tor_rtmock::MockSleepRuntime

pub struct MockSleepRuntime<R: Runtime> { /* private fields */ }

A wrapper Runtime that overrides the SleepProvider trait for the underlying runtime.

Implementations

impl<R: Runtime> MockSleepRuntime<R>

pub fn new(runtime: R) -> Self

Create a new runtime that wraps runtime, but overrides its view of time with a MockSleepProvider.

pub fn inner(&self) -> &R

Return a reference to the underlying runtime.

pub fn mock_sleep(&self) -> &MockSleepProvider

Return a reference to the MockSleepProvider

pub async fn advance(&self, dur: Duration)

See MockSleepProvider::advance()

pub fn jump_to(&self, new_wallclock: SystemTime)

See MockSleepProvider::jump_to()

pub fn wait_for<F: Future>(&self, fut: F) -> WaitFor<F>

Run a future under mock time, advancing time forward where necessary until it completes. Users of this function should read the whole of this documentation before using!

NOTE Instead of using this, consider MockRuntime, which will fully isolate the test case (albeit at the cost of demanding manual management of the simulated time).

The returned future will run fut, expecting it to create Sleeping futures (as returned by MockSleepProvider::sleep() and similar functions). When all such created futures have been polled (indicating the future is waiting on them), time will be advanced in order that the first (or only) of said futures returns Ready. This process then repeats until fut returns Ready itself (as in, the returned wrapper future will wait for all created Sleeping futures to be polled, and advance time again).

Note: The above described algorithm interacts poorly with futures that spawn asynchronous background tasks, or otherwise expect work to complete in the background before time is advanced. These futures will need to make use of the SleepProvider::block_advance (and similar) APIs in order to prevent time advancing while said tasks complete; see the documentation for those APIs for more detail.

Panics

Panics if another WaitFor future is already running. (If two ran simultaneously, they would both try and advance the same mock time clock, which would be bad.)