Struct tor_rtmock::time::MockSleepProvider

pub struct MockSleepProvider { /* private fields */ }

A dummy SleepProvider instance for testing.

The MockSleepProvider ignores the current time, and instead keeps its own view of the current Instant and SystemTime. You can advance them in-step by calling advance(), and you can simulate jumps in the system clock by calling jump().

This is not for production use.

Implementations

impl MockSleepProvider

pub fn new(wallclock: SystemTime) -> Self

Create a new MockSleepProvider, starting at a given wall-clock time.

pub async fn advance(&self, dur: Duration)

Advance the simulated timeline forward by dur.

Calling this function will wake any pending futures as appropriate, and yield to the scheduler so they get a chance to run.

Limitations

This function advances time in one big step. We might instead want to advance in small steps and make sure that each step’s futures can get run before the ones scheduled to run after it.

pub fn jump_to(&self, new_wallclock: SystemTime)

Simulate a discontinuity in the system clock, by jumping to new_wallclock.

Panics

Panics if we have already panicked while holding the lock on the internal timer state, and the lock is poisoned.